rename: 회원가입 폴더명 수정 및, boss, teacher 폴더 추가
