+++ app/src/main/java/com/example/teacherforboss/presentation/ui/auth/signup/basic/PasswordFragment.kt
@@ -1,10 +1,8 @@
-package com.example.teacherforboss.presentation.ui.auth.signup.fragment
+package com.example.teacherforboss.presentation.ui.auth.signup.basic
 
 import android.content.Context
-import android.content.Intent
 import android.os.Bundle
 import android.text.InputType
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
@@ -14,14 +12,11 @@ import android.widget.Toast
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
 import com.example.teacherforboss.R
 import com.example.teacherforboss.databinding.FragmentPasswordBinding
 import com.example.teacherforboss.presentation.ui.auth.signup.SignupActivity
 import com.example.teacherforboss.presentation.ui.auth.signup.SignupViewModel
 import com.example.teacherforboss.signup.fragment.NamePhoneFragment
-import dagger.hilt.android.AndroidEntryPoint
 
 //@AndroidEntryPoint
 class PasswordFragment : Fragment() {
