+++ app/src/main/res/layout/fragment_agreement.xml
@@ -3,7 +3,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.ui.auth.signup.fragment.AgreementFragment"
+    tools:context=".presentation.ui.auth.signup.basic.AgreementFragment"
     app:layout_behavior="com.google.android.material.bottomsheet.BottomSheetBehavior">
 
     <data>
